Got rid of sarcastic divide function, decided on doubles. It will be slower, which is a big deal for a chess heuristic, but this way it's more fine-grained. Integer division is almost useless since 50% of the time it will return 0.
Also, want to do function approximation as a toy example before moving
on the minichess.

+++ src/capps/gp/gptrees/GPFunction.java
@@ -11,7 +11,7 @@ package capps.gp.gptrees;
 public abstract class GPFunction extends GPNode {
 
 	@Override
-    public abstract int interpretForResult(GameState state); 
+    public abstract double interpretForResult(GameState state); 
 	@Override
     public abstract int numSubtrees(); 
 	@Override
